requirements.txt
discord.py>=2.3.2,<3.0.0
httpx>=0.27.0,<0.28.0
beautifulsoup4>=4.12.0,<5.0.0
openai>=1.12.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
aiohttp>=3.9.0,<4.0.0
